dmesg lexer refinements:
* move to textfmts.py where other logfiles are
* fix detection of dmesg -x (heuristic "unknown" state must be last)
* change styles for debug/warn lines
