+++ CHANGES
@@ -13,6 +13,7 @@ Version 2.6
 
 - Added lexers:
 
+  * Linux kernel logs (PR#1310)
   * Ride (PR#1319, PR#1321)
   * USD (PR#1290)
   * Mosel (PR#1287, PR#1326)
